// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  PATIENT: 'PATIENT',
  DRIVER: 'DRIVER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AmbulanceType = {
  BASIC: 'BASIC',
  ICU: 'ICU',
  FREEZER: 'FREEZER'
} as const

export type AmbulanceType = (typeof AmbulanceType)[keyof typeof AmbulanceType]


export const RequestPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL'
} as const

export type RequestPriority = (typeof RequestPriority)[keyof typeof RequestPriority]


export const DispatchStatus = {
  PENDING: 'PENDING',
  DISPATCHED: 'DISPATCHED',
  EN_ROUTE: 'EN_ROUTE',
  PICKED_UP: 'PICKED_UP',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type DispatchStatus = (typeof DispatchStatus)[keyof typeof DispatchStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]